Modified vm to make more sense
